Various improvements
Features:
Adding ability to restart application.
Adding application color palette support. CTRL-C in imageviewer.
Palette is persisted across sessions (tbd, reset to packaged defaults)
Adding find child by name recursive lookup.
Adding ability for dialog closure to return focus to previously held window by name.

Changes:
Rearranging the ImageViewer Control panel to only show "delete marked" button when viewing to be deleted filter.

Bug fixes:
Fixing bug in ARGB_to_AHSV not quantizing alpha correctly
Fixing AHSV_To_ARGB using the completely wrong function. yikes
Fixing bug in control panel not showing correct counts in filtered view buttons.

+++ ZFramework/ZBuffer.cpp
@@ -140,7 +140,7 @@ bool ZBuffer::LoadBuffer(const string& sFilename)
     }
 
     uint8_t* pScanFileData = (uint8_t*)ro_mmap.data();
-    uint64_t nFileSize = std::filesystem::file_size(sFilename);
+    uint32_t nFileSize = (uint32_t)std::filesystem::file_size(sFilename);
 
     mEXIF.clear();
     if (sExt == ".jpg" || sExt == ".jpeg")
